Replace perf monitoring with flamegraph instrumentation
