+++ Cargo.toml
@@ -15,7 +15,7 @@ thiserror = "1.0"
 hex = "0.4"
 parking_lot = "0.12"
 crossbeam-channel = "0.5"
-inferno = "0.11"
+flame = "0.2"
 ark-ff = "0.5"
 ark-ec = "0.5"
 ark-std = { version = "0.5", default-features = false, features = ["std"] }
@@ -33,12 +33,6 @@ bincode = "1.3"
 ff = "0.13"
 ctrlc = "3.4"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "time"] }
-
-[target.'cfg(windows)'.dependencies]
-windows = { version = "0.54", features = [
-    "Win32_Foundation",
-    "Win32_System_Threading",
-] }
 
 [features]
 default = []
