+++ src/main.rs
@@ -1,7 +1,7 @@
 // 引入 bpst 项目中的配置模块
 use bpst::config::{DeploymentConfig, P2PSimConfig};
 // 初始化日志系统
-use bpst::monitoring::perf::PerfExportGuard;
+use bpst::monitoring::flamegraph::FlamegraphExportGuard;
 use bpst::utils::init_logging;
 // 引入 bpst 项目中的 simulation 模块，包含运行节点、用户进程和P2P模拟的功能
 use bpst::simulation::{
@@ -11,7 +11,7 @@ use log::error;
 
 // Rust 程序的主入口函数
 fn main() {
-    let _perf_export_guard = PerfExportGuard::from_env();
+    let _flame_export_guard = FlamegraphExportGuard::from_env();
     // 初始化日志系统，确保控制台输出同时写入日志文件
     init_logging();
     // 获取命令行参数的迭代器
