+++ src/monitoring/mod.rs
@@ -1,1 +1,1 @@
-pub mod perf;
+pub mod flamegraph;
